Move cli parsing function to a separate file
