+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 pub mod trie;
 pub mod util;
 pub mod threads;
